requirements.txt
streamlit>=1.28.0
paddleocr>=2.7.0
paddlepaddle>=2.5.0
opencv-python>=4.8.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
configparser>=5.3.0
replicate>=0.15.0
requests>=2.31.0
python-dotenv>=1.0.0
regex>=2021.8.3
